+++ account-keeper-impl/src/main/java/com/dwarfeng/acckeeper/impl/bean/BeanMapper.java
@@ -26,7 +26,8 @@ import org.mapstruct.Mapping;
 @Mapper
 public interface BeanMapper {
 
-    // -----------------------------------------------------------Subgrade Key-----------------------------------------------------------
+    // region Subgrade Key
+
     HibernateLongIdKey longIdKeyToHibernate(LongIdKey longIdKey);
 
     @InheritInverseConfiguration
@@ -37,7 +38,10 @@ public interface BeanMapper {
     @InheritInverseConfiguration
     StringIdKey stringIdKeyFromHibernate(HibernateStringIdKey hibernateStringIdKey);
 
-    // -----------------------------------------------------------Acckeeper Key-----------------------------------------------------------
+    // endregion
+
+    // region Acckeeper Key
+
     HibernateProtectorVariableKey protectorVariableKeyToHibernate(ProtectorVariableKey protectorVariableKey);
 
     @InheritInverseConfiguration
@@ -48,7 +52,10 @@ public interface BeanMapper {
     @InheritInverseConfiguration
     RecordKey recordKeyFromHibernate(HibernateRecordKey hibernateRecordKey);
 
-    // -----------------------------------------------------------Acckeeper Entity-----------------------------------------------------------
+    // endregion
+
+    // region Acckeeper Entity
+
     @Mapping(target = "modifiedDatamark", ignore = true)
     @Mapping(target = "createdDatamark", ignore = true)
     @Mapping(target = "stringId", ignore = true)
@@ -120,4 +127,6 @@ public interface BeanMapper {
 
     @InheritInverseConfiguration
     DeriveHistory deriveHistoryFromHibernate(HibernateDeriveHistory hibernateDeriveHistory);
+
+    // endregion
 }